Problem solved completely added Instruction to run to ReadMe.md

+++ CurrencyConverter.java
@@ -1,4 +1,4 @@
 import java.rmi.*;
 public interface CurrencyConverter extends Remote{
-	public int converter(String from_currency, String to_currency,float amount)throws RemoteException;
+	public double converter(String from_currency, String to_currency,double amount)throws RemoteException;
 }
